Crate etl_core

This library provides a foundation to construct pipelines using several basic principles. There are stream producers and consumers, which must implement the traits crate::datastore::DataSource and crate::datastore::DataOutput. etl_job::job::JobRunner is used to construct pipelines using these elements. It also provides some simple state management which is loaded and saved using the crate::datastore::simple::SimpleStore trait which is designed for loading whole files It is important to note that the purpose of this library is to provide a framework to manage and organize pipelines. It is not meant to replace data anylitical tools or databases. The main use case for the author is performing 100 G JSON data merges and mapping them to a schema located on an MySql database.

Modules

batch

take in a stream and output same items but as batches

datastore

decoder

Used by the crate::datastore::DataSource to help decode various streams. So anything that implements a BytesSource target, can select which decoder they want to use. For example either crate::decoder::csv::CsvDecoder or crate::decoder::json::JsonDecoder

deps

deps which are re-exported (and used in the core)

encoder

joins

Perform joins between two crate::datastore::DataSources

preamble

queue

Adds facilities to implement queues

splitter

for splitting streams into many identical DataSources

task

transformer

utils