Trait etl_core::datastore::simple::SimpleStore

pub trait SimpleStore<T: Debug + 'static + Send>: Sync + Send {
    fn read_file_str<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<String, DataStoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn load<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<T, DataStoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn write<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        __arg1: &'life1 str, 
        __arg2: T
    ) -> Pin<Box<dyn Future<Output = Result<(), DataStoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait,
    { ... }
}

This is a simple store that acts like a key-val storage. It is not streamted so is not meant for big files. Primarily created for the JobRunner to store the state of the running job somewhere

Provided methods

fn read_file_str<'life0, 'life1, 'async_trait>(
&'life0 self,
__arg1: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<String, DataStoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

fn load<'life0, 'life1, 'async_trait>(
&'life0 self,
__arg1: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<T, DataStoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

fn write<'life0, 'life1, 'async_trait>(
&'life0 self,
__arg1: &'life1 str,
__arg2: T
) -> Pin<Box<dyn Future<Output = Result<(), DataStoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,