Trait etl_core::job::handler::StreamHandler

pub trait StreamHandler<T>: Sync + Send where
    T: DeserializeOwned + Serialize + Debug + Send + Sync, {
    fn shutdown<'life0, 'async_trait>(
        self: Box<Self>, 
        __arg1: &'life0 mut JobRunner
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn process_item<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        __arg1: JobItemInfo, 
        item: T, 
        job: &'life1 JobRunner
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn init<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        __arg1: &'life1 JobRunner
    ) -> Pin<Box<dyn Future<Output = Result<JobRunnerAction>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait,
    { ... }
}

Meant to be used for a variety situations like calling external apis, or reacting to the output of the stream in some fashion where returning a single element does not make sense

Required methods

fn shutdown<'life0, 'async_trait>(
self: Box<Self>,
__arg1: &'life0 mut JobRunner
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

fn process_item<'life0, 'life1, 'async_trait>(
&'life0 self,
__arg1: JobItemInfo,
item: T,
job: &'life1 JobRunner
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Provided methods

fn init<'life0, 'life1, 'async_trait>(
&'life0 mut self,
__arg1: &'life1 JobRunner
) -> Pin<Box<dyn Future<Output = Result<JobRunnerAction>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Optionally let the job handler decide if it needs to be skipped or resume. By default it will always start