Trait etl_core::queue::QueueClient

pub trait QueueClient<T: Debug + 'static + Send>: Sync + Send {
    fn start_incoming<'async_trait>(
        self: Box<Self>
    ) -> Pin<Box<dyn Future<Output = Result<DataSourceTask<T>, DataStoreError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
    { ... }

    fn pop<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Option<T>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn push<'life0, 'async_trait>(
        &'life0 self, 
        __arg1: T
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn pop_result<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Option<(T, Receiver<Result<T>>)>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }
}

Provided methods

fn start_incoming<'async_trait>(
self: Box<Self>
) -> Pin<Box<dyn Future<Output = Result<DataSourceTask<T>, DataStoreError>> + Send + 'async_trait>> where
Self: 'async_trait,

so any QueueClient can be turned into a DataSource

fn pop<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Option<T>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

fn push<'life0, 'async_trait>(
&'life0 self,
__arg1: T
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

fn pop_result<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Option<(T, Receiver<Result<T>>)>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

For cases when a produced item needs some action to be performed afterward by the queue, for example to perform some cleanup or deletion by the QueueClient implementation